Add files via upload
These are individual files for the biased data as well as the combined data that I will use for the project.  The original completely randomized data was (as expected) not choosing learning correlating variables.  The introduced bias in some of the numbers will hopefully introduce some pattern recognition that will allow me to see if I have built some models correctly.